+++ src/jvmTest/kotlin/choliver/nespot/cartridge/ChrMemoryTest.kt
@@ -2,8 +2,7 @@ package choliver.nespot.cartridge
 
 import choliver.nespot.*
 import choliver.nespot.cartridge.Rom.Mirroring.*
-import choliver.nespot.mappers.BankMappingChecker
-import choliver.nespot.mappers.BankMappingChecker.Companion.takesBytes
+import choliver.nespot.cartridge.BankMappingChecker.Companion.takesBytes
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import org.junit.jupiter.api.Nested
